+++ NWAT/NWAT/Views/Produktverwaltung_View.cs
@@ -1,20 +1,17 @@
-﻿using System;
+﻿using NWAT.DB;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using NWAT.DB;
 namespace NWAT
 {
     public partial class Produktverwaltung_View : Form
     {
         private ProductController prodCont;
-        public Produktverwaltung_View()
+        private Form parentView;
+        public Produktverwaltung_View(Form parentView)
         {
+            this.parentView = parentView;
             this.prodCont = new ProductController();
             InitializeComponent();
         }
@@ -38,6 +35,15 @@ namespace NWAT
             }
             
         }
+
+        /// <summary>
+        /// Handles the FormClosing event of the Productverwaltung_View control.
+        /// </summary>
+        /// <param name="sender">The source of the event.</param>
+        /// <param name="e">The <see cref="FormClosingEventArgs"/> instance containing the event data.</param>
+        /// Erstellt von Veit Berg, am 27.01.16
+        
+
         private void GetAllProdFromDB()
         {
 
@@ -54,7 +60,7 @@ namespace NWAT
             try{
             int selectedIndex = comboBox_ProdChoose.SelectedIndex;
             Product selectedItem = (Product)comboBox_ProdChoose.SelectedItem;
-            Product_Show_View ProdShow = new Product_Show_View(selectedItem.Product_Id);
+            Product_Show_View ProdShow = new Product_Show_View(this, selectedItem.Product_Id);
             ProdShow.Show();
             Hide();
             }
@@ -75,7 +81,7 @@ namespace NWAT
             try{
             int selectedIndex = comboBox_ProdChoose.SelectedIndex;
             Product selectedItem = (Product)comboBox_ProdChoose.SelectedItem;
-            Product_Update_View ProdUpdate = new Product_Update_View(selectedItem.Product_Id);
+            Product_Update_View ProdUpdate = new Product_Update_View(this, selectedItem.Product_Id);
             ProdUpdate.Show();
             Hide();
             }
@@ -93,15 +99,25 @@ namespace NWAT
         /// Erstellt von Veit Berg, am 27.01.16
         private void btn_ProdDelete_Click(object sender, EventArgs e)
         {
-            try{
-            int selectedIndex = comboBox_ProdChoose.SelectedIndex;
-            Product selectedItem = (Product)comboBox_ProdChoose.SelectedItem;
-            using (ProductController prodDelete = new ProductController()) {
+            try
+            {
+                int selectedIndex = comboBox_ProdChoose.SelectedIndex;
+                Product selectedItem = (Product)comboBox_ProdChoose.SelectedItem;
+                using (ProductController prodDelete = new ProductController())
+                {
 
-                prodDelete.DeleteProductFromDb(selectedItem.Product_Id);
-            }
-            
-            MessageBox.Show("Wollen Sie das ausgeählte Produkt wirklich löschen?");
+
+                    var result = MessageBox.Show("Wollen Sie das ausgeählte Produkt wirklich löschen?",
+                        "Löschbestätigung",
+                        MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                    if (result == DialogResult.Yes)
+                    {
+                        prodDelete.DeleteProductFromDb(selectedItem.Product_Id);
+                        RefreshDropdown();
+                    }
+                }
+
             }
             catch (Exception x)
             {
@@ -122,7 +138,7 @@ namespace NWAT
         private void btn_ProdCreate_Click(object sender, EventArgs e)
         {
             try{
-            Product_Create_View ProdCreate = new Product_Create_View();
+            Product_Create_View ProdCreate = new Product_Create_View(this);
             ProdCreate.Show();
             Hide();
             }
@@ -140,21 +156,44 @@ namespace NWAT
         /// Erstellt von Veit Berg, am 27.01.16
         private void btn_refresh_Click(object sender, EventArgs e)
         {
-           try{
-            using (ProductController ProdRefr = new ProductController())
+            RefreshDropdown();
+        }
+
+        /// <summary>
+        /// refreshs dropdown menue
+        /// </summary>
+        /// <param name="sender">The source of the event.</param>
+        /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
+        /// Erstellt von Veit Berg, am 27.01.16
+        public void RefreshDropdown()
+        {
+            try
             {
-                List<Product> ProdList = ProdRefr.GetAllProductsFromDb();
-                var bindingList = new BindingList<Product>(ProdList);
-                var source = new BindingSource(bindingList, null);
-                comboBox_ProdChoose.DataSource = ProdList;
-                comboBox_ProdChoose.DisplayMember = "Name";
-                comboBox_ProdChoose.ValueMember = "Product_ID";
+                using (ProductController ProdRefr = new ProductController())
+                {
+                    List<Product> ProdList = ProdRefr.GetAllProductsFromDb();
+                    var bindingList = new BindingList<Product>(ProdList);
+                    var source = new BindingSource(bindingList, null);
+                    comboBox_ProdChoose.DataSource = ProdList;
+                    comboBox_ProdChoose.DisplayMember = "Name";
+                    comboBox_ProdChoose.ValueMember = "Product_ID";
+                }
             }
-           }
-           catch (Exception x)
-           {
-               MessageBox.Show("Ups da lief was schief");
-           }
+            catch (Exception x)
+            {
+                MessageBox.Show("Ups da lief was schief");
+            }
+        }
+
+        /// <summary>
+        /// Handles the FormClosing event of the Productverwaltung_View control.
+        /// </summary>
+        /// <param name="sender">The source of the event.</param>
+        /// <param name="e">The <see cref="FormClosingEventArgs"/> instance containing the event data.</param>
+        /// Erstellt von Veit Berg, am 27.01.16
+        private void Produktverwaltung_FormClosing(object sender, FormClosingEventArgs e)
+        {
+            this.parentView.Show();
         }
     }
     public class aktRowProd
